Remove FakePackage from Method
This value was being copied from Fake for the sole purpose of rendering
a template. The FakePackage field does not make much sense as a field on a
Method struct, it is not a property of a method.

Instead of copying the value to a struct, change the templates to
reference the root object using $, which is always the initial value of
'.' when a template rendering is executed.