refactor arguments package
Use error as return values instead of calling fail. Calling fail to
os.Exit makes code difficult to test, and will not allow defers to run
when there is an error.

Remove unnecessary interfaces. Use a single function shim for handling
paths.

Remove SourcePackageDir field, it is not read outside of tests.